Add FET integration test, remove unused code.
