+++ Timetabling.Tests/Objects/Constraints/TimeConstraints/ConstraintActivitiesPreferredStartingTimesTest.cs
@@ -1,27 +1,15 @@
-﻿using NUnit.Framework;
-using System.Linq;
-using Moq;
-using Timetabling.DB;
-using System.Collections.Generic;
-using System.Data.Entity;
-using Timetabling.Objects.Constraints.TimeConstraints;
+﻿using System.Linq;
+using NUnit.Framework;
 using Timetabling.Objects;
+using Timetabling.Objects.Constraints.TimeConstraints;
 
-namespace Timetabling.Tests.Objects.Constraints.TimeConstraints.Tests
+namespace Timetabling.Tests.Objects.Constraints.TimeConstraints
 {
-    [TestFixture()]
+
     internal class ConstraintActivitiesPreferredStartingTimesTest
     {
-        Mock<DataModel> test;
 
-        [SetUp]
-        public void Init()
-        {
-            test = new TestDataModel().MockDataModel;
-
-        }
-
-        [Test()]
+        [Test]
         public void TestConstruct()
 
         {
@@ -37,7 +25,7 @@ namespace Timetabling.Tests.Objects.Constraints.TimeConstraints.Tests
             var constraintTest = new ConstraintActivitiesPreferredStartingTimes { DaysList = { (Days)2 }, SubjectId = 1, HoursList = { 3 }, NumberOfHours = 1 };
             var constraintTest2 = new ConstraintActivitiesPreferredStartingTimes { DaysList = { (Days)3, (Days)3, (Days)2 }, SubjectId = 1, HoursList = { 3, 4, 5 }, NumberOfHours = 3 };
 
-            var result = constraint.Create(test.Object);
+            var result = constraint.Create(new TestDataModel().MockDataModel.Object);
 
             Assert.AreEqual(1, result.Count(item => item.ToString().Equals(constraintTest.ToXelement().ToString())));
             Assert.AreEqual(0, result.Count(item => item.ToString().Equals(constraintTest2.ToXelement().ToString())));
