+++ Timetabling.Tests/Objects/SubjectsListTest.cs
@@ -48,8 +48,7 @@ namespace Timetabling.Tests.Objects
         [Test]
         public void SubjectNotActive()
         {
-            Assert.AreEqual(0, test.Elements("Subject").Elements("Name").Count(item => item.Value.Equals("0")));
-
+            Assert.AreEqual(0, test.Elements("Subject").Elements("Name").Count(item => item.Value.Equals("2")));
         }
     }
 }
